+++ packages/lito/src/Browse/EditorialElement.tsx
@@ -1,0 +1,16 @@
+import { ResourceList } from '../components'
+
+export interface EditorialElementProps {
+  value: EditorialElement
+}
+
+export interface EditorialElement {
+  name?: string
+  contents: any[]
+}
+
+const EditorialElement = ({ value }: EditorialElementProps) => {
+  return <ResourceList title={value.name} resources={value.contents} />
+}
+
+export default EditorialElement
