+++ packages/lito/src/api/fetcher.ts
@@ -1,0 +1,8 @@
+export const fetcher = async (url: string) => {
+  const {
+    api,
+    _bag: { store },
+  } = MusicKit.getInstance() as any
+  const { data } = await api.music(url.replace(/\{storefrontId\}/g, api.storefrontId))
+  return store.populateDataRecords(data)
+}
