+++ packages/lito/src/components/Header.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const Header = styled.div`
+  margin: 32px 40px 16px;
+  font-size: 34px;
+  font-weight: 500;
+`
+
+export const SubHeader = styled.div`
+  padding: 13px 0;
+  font-size: 17px;
+  line-height: 1.29412;
+`
